重构评分清洗任务流
